requirements.txt
# This file contains all the development requirements for our linting toolchain.
# Don't forget to pin your dependencies!
# This list will have to be migrated if you wish to use another dependency manager.

asyncpg == 0.29.0
asyncpg-stubs == 0.29.1
disnake[speed]==2.9.2
python-dotenv==1.0.1
ruff~=0.5.0
pre-commit~=3.7.1
cfgv==3.4.0
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
mypy==1.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
pyyaml==6.0.1
ruff==0.5.2
typing-extensions==4.12.2
virtualenv==20.26.3
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
asyncpg-stubs==0.29.1
attrs==23.2.0
brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
disnake==2.9.2
distlib==0.3.8
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
multidict==6.0.5
mypy==1.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.6
platformdirs==4.2.2
pre-commit==3.7.1
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
pyyaml==6.0.1
ruff==0.5.2
sniffio==1.3.1
typing-extensions==4.12.2
virtualenv==20.26.3
yarl==1.9.4
